% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-model-methods.R
\name{as_mcmc.mixgpd_fit}
\alias{as_mcmc.mixgpd_fit}
\title{As mcmc mixgpd fit}
\usage{
\method{as_mcmc}{mixgpd_fit}(object, ...)
}
\arguments{
\item{object}{object.}

\item{...}{Additional arguments.}
}
\value{
See details.
}
\description{
As mcmc mixgpd fit.
}
\examples{
f <- getFromNamespace("as_mcmc.mixgpd_fit", "DPmixGPD")
f
\dontrun{
# Example usage (may require MCMC run and nimble compilation).
}

}
