% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-model-methods.R
\name{as_mcmc}
\alias{as_mcmc}
\title{As mcmc}
\usage{
as_mcmc(object, ...)
}
\arguments{
\item{object}{object.}

\item{...}{Additional arguments.}
}
\value{
See details.
}
\description{
As mcmc.
}
\examples{
f <- getFromNamespace("as_mcmc", "DPmixGPD")
f
\dontrun{
# Example usage (may require MCMC run and nimble compilation).
}

}
