% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelspec-dp.R
\name{build_model_spec}
\alias{build_model_spec}
\title{Build model spec}
\usage{
build_model_spec(
  Y,
  X,
  kernel,
  tail,
  dp_rep,
  priors = list(),
  dp_ctrl = list(),
  trans = list(),
  alpha = 0.05
)
}
\arguments{
\item{Y}{Y.}

\item{X}{X.}

\item{kernel}{kernel.}

\item{tail}{tail.}

\item{dp_rep}{dp_rep.}

\item{priors}{priors.}

\item{dp_ctrl}{dp_ctrl.}

\item{trans}{trans.}

\item{alpha}{alpha.}
}
\value{
See details.
}
\description{
Build model spec.
}
\examples{
f <- getFromNamespace("build_model_spec", "DPmixGPD")
f

}
\keyword{internal}
