% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelspec-dp.R
\name{build_model_spec_xy}
\alias{build_model_spec_xy}
\title{Build model spec xy}
\usage{
build_model_spec_xy(
  Y,
  X = NULL,
  mode = c("response_only", "regression"),
  kernel = c("gamma", "lognormal", "normal", "laplace", "inverse_gaussian", "amoroso",
    "pareto"),
  dp_rep = c("stick_breaking", "crp"),
  dp_ctrl = list(),
  priors = list(),
  trans = NULL,
  tail = c("none", "gpd")
)
}
\arguments{
\item{Y}{Y.}

\item{X}{X.}

\item{mode}{mode.}

\item{kernel}{kernel.}

\item{dp_rep}{dp_rep.}

\item{dp_ctrl}{dp_ctrl.}

\item{priors}{priors.}

\item{trans}{trans.}

\item{tail}{tail.}
}
\value{
See details.
}
\description{
Build model spec xy.
}
\examples{
f <- getFromNamespace("build_model_spec_xy", "DPmixGPD")
f

}
\keyword{internal}
