% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-engine.R
\name{build_nimble_model_gamma_reg}
\alias{build_nimble_model_gamma_reg}
\title{Internal: run Nimble MCMC for unconditional Gamma SB mixture}
\usage{
build_nimble_model_gamma_reg(spec)
}
\arguments{
\item{spec}{spec.}
}
\value{
See details.
}
\description{
Internal: run Nimble MCMC for unconditional Gamma SB mixture
}
\examples{
f <- getFromNamespace("build_nimble_model_gamma_reg", "DPmixGPD")
f

}
\keyword{Build}
\keyword{gamma}
\keyword{internal}
\keyword{model}
\keyword{nimble}
\keyword{reg}
\keyword{reg.}
