% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_gamma_nimble.R
\name{build_nimble_model_gamma_uncond}
\alias{build_nimble_model_gamma_uncond}
\title{Internal: build Nimble model for unconditional Gamma SB mixture}
\usage{
build_nimble_model_gamma_uncond(spec)
}
\description{
Internal: build Nimble model for unconditional Gamma SB mixture
}
\keyword{internal}
