% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-build-and-run.R
\name{check_dpmixgpd_bundle}
\alias{check_dpmixgpd_bundle}
\title{Validate a dpmixgpd_bundle before running MCMC}
\usage{
check_dpmixgpd_bundle(bundle, strict = FALSE)
}
\arguments{
\item{bundle}{A \code{dpmixgpd_bundle}.}

\item{strict}{Logical; if TRUE, stop() on errors.}
}
\value{
A list with $ok, $errors, $warnings, and some key derived info.
}
\description{
Validate a dpmixgpd_bundle before running MCMC
}
\keyword{internal}
