% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncore-utils.R
\name{.apply_trans}
\alias{.apply_trans}
\title{Apply a link/transform}
\usage{
.apply_trans(x, trans)
}
\arguments{
\item{x}{Numeric vector.}

\item{trans}{Character scalar: "identity" or "exp".}
}
\value{
Numeric vector.
}
\description{
Apply a link/transform
}
\keyword{internal}
