% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_effect.R
\name{.as_mcmc_matrix}
\alias{.as_mcmc_matrix}
\title{Coerce MCMC draws to a matrix (iterations x parameters)}
\usage{
.as_mcmc_matrix(x)
}
\description{
Coerce MCMC draws to a matrix (iterations x parameters)
}
\keyword{internal}
