% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelspec-dp.R
\name{.bracket_root}
\alias{.bracket_root}
\title{Bracket root}
\usage{
.bracket_root(
  fun,
  target = 0,
  lower = 0,
  upper = 1,
  expand = 2,
  max_expand = 60
)
}
\arguments{
\item{fun}{fun.}

\item{target}{target.}

\item{lower}{lower.}

\item{upper}{upper.}

\item{expand}{expand.}

\item{max_expand}{max_expand.}
}
\value{
See details.
}
\description{
Bracket root.
}
\examples{
f <- getFromNamespace(".bracket_root", "DPmixGPD")
f

}
\keyword{internal}
