% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{.check_dp_ctrl}
\alias{.check_dp_ctrl}
\title{Check dp ctrl}
\usage{
.check_dp_ctrl(dp_ctrl, N)
}
\arguments{
\item{dp_ctrl}{dp_ctrl.}

\item{N}{N.}
}
\value{
See details.
}
\description{
Check dp ctrl.
}
\examples{
f <- getFromNamespace(".check_dp_ctrl", "DPmixGPD")
f

}
\keyword{internal}
