% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncore-utils.R
\name{.check_formula_data}
\alias{.check_formula_data}
\title{Check formula and data}
\usage{
.check_formula_data(formula, data)
}
\arguments{
\item{formula}{A model formula with a numeric response.}

\item{data}{A data.frame containing variables in \code{formula}.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Check formula and data
}
\keyword{internal}
