% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{.check_kernel_support}
\alias{.check_kernel_support}
\title{Check kernel support}
\usage{
.check_kernel_support(Y, kernel)
}
\arguments{
\item{Y}{Y.}

\item{kernel}{kernel.}
}
\value{
See details.
}
\description{
Check kernel support.
}
\examples{
f <- getFromNamespace(".check_kernel_support", "DPmixGPD")
f

}
\keyword{internal}
