% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{.check_mcmc}
\alias{.check_mcmc}
\title{Check mcmc}
\usage{
.check_mcmc(mcmc)
}
\arguments{
\item{mcmc}{mcmc.}
}
\value{
See details.
}
\description{
Check mcmc.
}
\examples{
f <- getFromNamespace(".check_mcmc", "DPmixGPD")
f

}
\keyword{internal}
