% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{.check_missing}
\alias{.check_missing}
\title{Check missing}
\usage{
.check_missing(Y, X)
}
\arguments{
\item{Y}{Y.}

\item{X}{X.}
}
\value{
See details.
}
\description{
Check missing.
}
\examples{
f <- getFromNamespace(".check_missing", "DPmixGPD")
f

}
\keyword{internal}
