% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-utilities-internal.R
\name{.compute_ess}
\alias{.compute_ess}
\title{Compute ESS if coda is installed (otherwise NA)}
\usage{
.compute_ess(draws)
}
\arguments{
\item{mat}{Draw matrix (iter x p).}
}
\value{
Named numeric vector of ESS (length p).
}
\description{
Compute ESS if coda is installed (otherwise NA)
}
\keyword{internal}
