% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-engine-crp-gamma.R
\name{.crp_gamma_logprior}
\alias{.crp_gamma_logprior}
\title{Log prior density for (shape, scale) under H (lognormal).}
\usage{
.crp_gamma_logprior(shape, scale, pri)
}
\description{
Log prior density for (shape, scale) under H (lognormal).
}
\keyword{internal}
