% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-engine-crp-gamma.R
\name{.crp_gamma_priors}
\alias{.crp_gamma_priors}
\title{Extract/default priors for CRP Gamma engine (Option B: log-params normal).}
\usage{
.crp_gamma_priors(spec)
}
\description{
Extract/default priors for CRP Gamma engine (Option B: log-params normal).
}
\keyword{internal}
