% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-engine-crp-gamma.R
\name{.crp_update_alpha}
\alias{.crp_update_alpha}
\title{Escobar-West update for alpha}
\usage{
.crp_update_alpha(alpha, K, N, a0, b0)
}
\description{
Escobar-West update for alpha
}
\keyword{internal}
