% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-utilities-internal.R
\name{.crp_weights_from_z}
\alias{.crp_weights_from_z}
\title{Compute CRP weights from a vector of allocations z}
\usage{
.crp_weights_from_z(z, Kmax)
}
\arguments{
\item{z}{Integer vector of length N with values in 1:Kmax.}

\item{Kmax}{Integer.}
}
\value{
Numeric vector of length Kmax summing to 1.
}
\description{
Compute CRP weights from a vector of allocations z
}
\keyword{internal}
