% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{.ensure_nimble_compat}
\alias{.ensure_nimble_compat}
\title{Ensure nimble is available}
\usage{
.ensure_nimble_compat()
}
\description{
Internal guard used before building/running nimble models.
}
\keyword{internal}
