% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_effect.R
\name{.extract_gamma_dp_params}
\alias{.extract_gamma_dp_params}
\title{Extract ordered Gamma DP mixture parameters from a draw-matrix
Returns W, Shape, Scale as matrices M x K}
\usage{
.extract_gamma_dp_params(draws, renormalize_weights = TRUE)
}
\description{
Extract ordered Gamma DP mixture parameters from a draw-matrix
Returns W, Shape, Scale as matrices M x K
}
\keyword{internal}
