% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-model-methods.R
\name{.extract_gamma_dp_params_uncond}
\alias{.extract_gamma_dp_params_uncond}
\title{Extract gamma dp params uncond.}
\usage{
.extract_gamma_dp_params_uncond(object)
}
\arguments{
\item{object}{object.}
}
\value{
See details.
}
\description{
Extract gamma dp params uncond.
}
\examples{
f <- getFromNamespace(".extract_gamma_dp_params_uncond", "DPmixGPD")
f
}
\keyword{internal}
