% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-engine-sb-gamma.R
\name{.extract_sb_gamma_reg_params}
\alias{.extract_sb_gamma_reg_params}
\title{Extract sb gamma reg params}
\usage{
.extract_sb_gamma_reg_params(draws, K, p, renormalize_weights = TRUE)
}
\description{
Extract sb gamma reg params
}
\keyword{internal}
