% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-te-fit.R
\name{.fit_propensity}
\alias{.fit_propensity}
\title{Fit propensity}
\usage{
.fit_propensity(A, X, method = c("glm_logit"), ...)
}
\arguments{
\item{A}{A.}

\item{X}{X.}

\item{method}{method.}

\item{...}{Additional arguments.}
}
\value{
See details.
}
\description{
Fit propensity.
}
\examples{
f <- getFromNamespace(".fit_propensity", "DPmixGPD")
f

}
\keyword{internal}
