% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_effect.R
\name{.gamma_mix_cdf_1draw}
\alias{.gamma_mix_cdf_1draw}
\title{Mixture CDF for one draw (Gamma mixture)}
\usage{
.gamma_mix_cdf_1draw(x, W, Shape, Scale)
}
\description{
Mixture CDF for one draw (Gamma mixture)
}
\keyword{internal}
