% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-utilities-internal.R
\name{.get_mix_fns}
\alias{.get_mix_fns}
\title{Resolve the four MIX functions (d/p/q/r) for a kernel + tail flag}
\usage{
.get_mix_fns(kernel, GPD)
}
\description{
Resolve the four MIX functions (d/p/q/r) for a kernel + tail flag
}
\keyword{internal}
