% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-model-methods.R
\name{.invert_cdf_uniroot}
\alias{.invert_cdf_uniroot}
\title{Invert cdf uniroot}
\usage{
.invert_cdf_uniroot(
  cdf_fun,
  tau,
  lower = 0,
  upper = NULL,
  max_expand = 60L,
  expand_factor = 2,
  tol = 1e-08
)
}
\arguments{
\item{cdf_fun}{cdf_fun.}

\item{tau}{tau.}

\item{lower}{lower.}

\item{upper}{upper.}

\item{max_expand}{max_expand.}

\item{expand_factor}{expand_factor.}

\item{tol}{tol.}
}
\value{
See details.
}
\description{
Invert cdf uniroot.
}
\examples{
f <- getFromNamespace(".invert_cdf_uniroot", "DPmixGPD")
f

}
\keyword{internal}
