% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelspec-dp.R
\name{.mix_cdf_gamma}
\alias{.mix_cdf_gamma}
\title{Mix cdf gamma}
\usage{
.mix_cdf_gamma(x, w, shape, scale)
}
\arguments{
\item{x}{x.}

\item{w}{w.}

\item{shape}{shape.}

\item{scale}{scale.}
}
\value{
See details.
}
\description{
Mix cdf gamma.
}
\examples{
f <- getFromNamespace(".mix_cdf_gamma", "DPmixGPD")
f

}
\keyword{internal}
