% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelspec-dp.R
\name{.normalize_weights}
\alias{.normalize_weights}
\title{Normalize weights}
\usage{
.normalize_weights(w)
}
\arguments{
\item{w}{w.}
}
\value{
See details.
}
\description{
Normalize weights.
}
\examples{
f <- getFromNamespace(".normalize_weights", "DPmixGPD")
f

}
\keyword{internal}
