% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-utilities-internal.R
\name{.plot_dispatch}
\alias{.plot_dispatch}
\title{Dispatch plotting for common MCMC diagnostics}
\usage{
.plot_dispatch(mat, family, params, ...)
}
\arguments{
\item{mat}{Draw matrix.}

\item{family}{One of trace/density/acf/pairs.}

\item{params}{Parameters to plot.}
}
\description{
Dispatch plotting for common MCMC diagnostics
}
\keyword{internal}
