% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-engine.R
\name{.pp_cdf_gamma_uncond}
\alias{.pp_cdf_gamma_uncond}
\title{Pp cdf gamma uncond}
\usage{
.pp_cdf_gamma_uncond(params, x)
}
\arguments{
\item{params}{params.}

\item{x}{x.}
}
\value{
See details.
}
\description{
Pp cdf gamma uncond.
}
\examples{
f <- getFromNamespace(".pp_cdf_gamma_uncond", "DPmixGPD")
f

}
\keyword{internal}
