% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncore-utils.R
\name{.prepare_design_matrix}
\alias{.prepare_design_matrix}
\title{Prepare design matrix}
\usage{
.prepare_design_matrix(formula, data)
}
\arguments{
\item{formula}{Model formula.}

\item{data}{Data frame.}
}
\value{
A list with \code{y}, \code{X}, \code{has_regression}, and \code{terms}.
}
\description{
The package treats models as unconditional if the right-hand side of the
formula is empty (for example \code{y ~ 1}).
}
\keyword{internal}
