% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelspec-dp.R
\name{.q_gamma_dp_one_draw}
\alias{.q_gamma_dp_one_draw}
\title{Q gamma dp one draw}
\usage{
.q_gamma_dp_one_draw(
  p,
  w,
  shape,
  scale,
  lower = 0,
  upper0 = 1,
  expand = 2,
  max_expand = 60,
  tol = .Machine$double.eps^0.5,
  maxiter = 1000
)
}
\arguments{
\item{p}{p.}

\item{w}{w.}

\item{shape}{shape.}

\item{scale}{scale.}

\item{lower}{lower.}

\item{upper0}{upper0.}

\item{expand}{expand.}

\item{max_expand}{max_expand.}

\item{tol}{tol.}

\item{maxiter}{maxiter.}
}
\value{
See details.
}
\description{
Q gamma dp one draw.
}
\examples{
f <- getFromNamespace(".q_gamma_dp_one_draw", "DPmixGPD")
f

}
\keyword{internal}
