% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-quantile-gamma-dp.R
\name{.q_gamma_dp_over_draws}
\alias{.q_gamma_dp_over_draws}
\title{Vectorized quantiles over posterior draws (returns M x length(p))}
\usage{
.q_gamma_dp_over_draws(
  p_vec,
  W,
  Shape,
  Scale,
  lower = 0,
  upper0 = 1,
  expand = 2,
  max_expand = 60,
  tol = .Machine$double.eps^0.5,
  maxiter = 1000
)
}
\description{
Vectorized quantiles over posterior draws (returns M x length(p))
}
\keyword{internal}
