% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_effect.R
\name{.qgamma_mix_1draw}
\alias{.qgamma_mix_1draw}
\title{Robust uniroot quantile solver for one draw (Gamma mixture)
Uses increasing upper bound until CDF(upper) >= p}
\usage{
.qgamma_mix_1draw(p, W, Shape, Scale, y_max_hint = 1)
}
\description{
Robust uniroot quantile solver for one draw (Gamma mixture)
Uses increasing upper bound until CDF(upper) >= p
}
\keyword{internal}
