% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelspec-dp.R
\name{.qmix_gamma_dp_uncond}
\alias{.qmix_gamma_dp_uncond}
\title{Qmix gamma dp uncond.}
\usage{
.qmix_gamma_dp_uncond(
  tau,
  w,
  shape,
  scale,
  lower = 0,
  upper = NULL,
  expand = 2,
  max_expand = 60,
  tol = 1e-10,
  maxiter = 1000
)
}
\arguments{
\item{tau}{tau.}

\item{w}{w.}

\item{shape}{shape.}

\item{scale}{scale.}

\item{lower}{lower.}

\item{upper}{upper.}

\item{expand}{expand.}

\item{max_expand}{max_expand.}

\item{tol}{tol.}

\item{maxiter}{maxiter.}
}
\value{
See details.
}
\description{
Qmix gamma dp uncond.
}
\examples{
f <- getFromNamespace(".qmix_gamma_dp_uncond", "DPmixGPD")
f

}
\keyword{internal}
