% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncore-utils.R
\name{.summarize_draw_matrix}
\alias{.summarize_draw_matrix}
\title{Summarize draws to a tidy matrix}
\usage{
.summarize_draw_matrix(draws, prob = c(0.025, 0.975))
}
\arguments{
\item{draws}{Numeric matrix with rows as draws.}

\item{prob}{Numeric vector of length 2 giving lower/upper credible probs.}
}
\value{
A matrix with columns mean, sd, q_low, q_high.
}
\description{
Summarize draws to a tidy matrix
}
\keyword{internal}
