% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-model-methods.R
\name{.summarize_mcmc}
\alias{.summarize_mcmc}
\title{Summarize mcmc}
\usage{
.summarize_mcmc(mcmc_draws, alpha = 0.05)
}
\arguments{
\item{mcmc_draws}{mcmc_draws.}

\item{alpha}{alpha.}
}
\value{
See details.
}
\description{
Summarize mcmc.
}
\examples{
f <- getFromNamespace(".summarize_mcmc", "DPmixGPD")
f

}
\keyword{internal}
