% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{.transform_generate_code}
\alias{.transform_generate_code}
\title{Transform generate code}
\usage{
.transform_generate_code(lhs, lp, transform)
}
\arguments{
\item{lhs}{lhs.}

\item{lp}{lp.}

\item{transform}{transform.}
}
\value{
See details.
}
\description{
Transform generate code.
}
\examples{
f <- getFromNamespace(".transform_generate_code", "DPmixGPD")
f

}
\keyword{internal}
