% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{.transform_resolve}
\alias{.transform_resolve}
\title{Transform resolve}
\usage{
.transform_resolve(X, trans, caller = "fit.dpm")
}
\arguments{
\item{X}{X.}

\item{trans}{trans.}

\item{caller}{caller.}
}
\value{
See details.
}
\description{
Transform resolve.
}
\examples{
f <- getFromNamespace(".transform_resolve", "DPmixGPD")
f

}
\keyword{internal}
