% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{.transform_valid_names}
\alias{.transform_valid_names}
\title{Transform valid names}
\usage{
.transform_valid_names(scale_spec, X, caller = ".transform_resolve")
}
\arguments{
\item{scale_spec}{scale_spec.}

\item{X}{X.}

\item{caller}{caller.}
}
\value{
See details.
}
\description{
Transform valid names.
}
\examples{
f <- getFromNamespace(".transform_valid_names", "DPmixGPD")
f

}
\keyword{internal}
