% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\docType{data}
\name{.valid_trans_labels}
\alias{.valid_trans_labels}
\title{Internal constants}
\format{
An object of class \code{character} of length 4.
}
\usage{
.valid_trans_labels
}
\value{
See details.
}
\description{
Internal constants used by the package.
}
\examples{
getFromNamespace(".valid_trans_labels", "DPmixGPD")

}
\keyword{internal}
