% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelspec-dp.R
\name{.validate_trans}
\alias{.validate_trans}
\title{Validate trans}
\usage{
.validate_trans(kinfo, trans)
}
\arguments{
\item{kinfo}{kinfo.}

\item{trans}{trans.}
}
\value{
See details.
}
\description{
Validate trans.
}
\examples{
f <- getFromNamespace(".validate_trans", "DPmixGPD")
f

}
\keyword{internal}
