% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-utilities-internal.R
\name{.validate_X_pred}
\alias{.validate_X_pred}
\title{Validate/standardize prediction X (design) against training X}
\usage{
.validate_X_pred(Xpred, Xtrain)
}
\description{
Validate/standardize prediction X (design) against training X
}
\keyword{internal}
