% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.wrap_scalar_first_arg}
\alias{.wrap_scalar_first_arg}
\title{Wrap scalar first-argument functions to handle vector inputs.}
\usage{
.wrap_scalar_first_arg(fun, first_arg_name)
}
\description{
Wrap scalar first-argument functions to handle vector inputs.
}
\details{
Many low-level distribution helpers are scalar in their first argument. This
wrapper lifts such functions to vector inputs by evaluating the scalar
function repeatedly and combining the results into either a numeric vector or
a matrix, depending on the length of the original return value.
}
\keyword{internal}
