% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\docType{package}
\name{DPmixGPD-package}
\alias{DPmixGPD}
\alias{DPmixGPD-package}
\title{DPmixGPD: Dirichlet process mixtures with optional GPD tail}
\description{
An experimental framework for Dirichlet process mixture models (including
regression and unconditional formulations) with flexible kernels and an
optional generalized Pareto (GPD) tail extension.
}
\author{
\strong{Maintainer}: Arnab Aich \email{arnab@example.com}

}
\keyword{internal}
