% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPmixGPD-package.R
\docType{package}
\name{DPmixGPD}
\alias{DPmixGPD}
\alias{DPmixGPD-package}
\title{DPmixGPD: Dirichlet Process Mixtures with Generalized Pareto Tail}
\description{
\code{DPmixGPD} provides Bayesian Dirichlet process mixture models for
unconditional, conditional, and causal workflows, with optional generalized
Pareto tail modeling.
}
\details{
Core entry points include \code{\link[=build_nimble_bundle]{build_nimble_bundle()}}, \code{\link[=run_mcmc_bundle_manual]{run_mcmc_bundle_manual()}},
\code{\link[=build_causal_bundle]{build_causal_bundle()}}, and \code{\link[=run_mcmc_causal]{run_mcmc_causal()}}.

Learn more via:
\itemize{
\item package index: \code{help(package = "DPmixGPD")}
\item vignettes: \code{browseVignettes("DPmixGPD")}
\item vignette pages:
\itemize{
\item \code{vignette("01-basic", package = "DPmixGPD")}
\item \code{vignette("03-unconditional", package = "DPmixGPD")}
\item \code{vignette("04-conditional", package = "DPmixGPD")}
\item \code{vignette("05-causal", package = "DPmixGPD")}
}
\item installed HTML pages: \href{../doc/01-basic.html}{01-basic.html},
\href{../doc/03-unconditional.html}{03-unconditional.html},
\href{../doc/04-conditional.html}{04-conditional.html},
\href{../doc/05-causal.html}{05-causal.html}
\item website: \url{https://arnabaich96.github.io/DPmixGPD/}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://arnabaich96.github.io/DPmixGPD/}
  \item \url{https://arnabaich96.github.io/DPmixGPD/pkgdown/}
  \item Report bugs at \url{https://github.com/arnabaich96/DPmixGPD/issues}
}

}
\author{
\strong{Maintainer}: Arnab Aich \email{aaich@fsu.edu}

}
