% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dpm_wrappers.R
\name{fit.dpm}
\alias{fit.dpm}
\title{Fit a Dirichlet process mixture model (no tail)}
\usage{
fit.dpm(
  formula,
  data,
  kernel = c("gamma", "lognormal", "normal", "laplace", "inverse_gaussian", "amoroso",
    "pareto"),
  dp_rep = c("stick_breaking", "crp"),
  dp_ctrl = list(K = 5),
  priors = list(),
  mcmc = list(),
  alpha = 0.05,
  trans = list(),
  intercept = FALSE
)
}
\description{
Fit a Dirichlet process mixture model (no tail)
}
