% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-te-fit.R
\name{fit.TE}
\alias{fit.TE}
\title{Fit TE}
\usage{
fit.TE(
  formula,
  data,
  A,
  kernel = c("gamma", "lognormal", "normal", "laplace", "inverse_gaussian", "amoroso",
    "pareto"),
  tail = c("none", "gpd"),
  dp_rep = c("stick_breaking", "crp"),
  dp_ctrl = list(K = 5),
  priors = list(),
  mcmc = list(),
  alpha = 0.05,
  trans = list(),
  intercept = TRUE
)
}
\arguments{
\item{formula}{formula.}

\item{data}{data.}

\item{A}{A.}

\item{kernel}{kernel.}

\item{tail}{tail.}

\item{dp_rep}{dp_rep.}

\item{dp_ctrl}{dp_ctrl.}

\item{priors}{priors.}

\item{mcmc}{mcmc.}

\item{alpha}{alpha.}

\item{trans}{trans.}

\item{intercept}{intercept.}
}
\value{
See details.
}
\description{
Fit TE.
}
\examples{
f <- getFromNamespace("fit.TE", "DPmixGPD")
f
\dontrun{
# Example usage (may require MCMC run and nimble compilation).
}

}
