% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_registry.R
\name{get_kernel}
\alias{get_kernel}
\title{Get a registered kernel}
\usage{
get_kernel(kernel)
}
\arguments{
\item{kernel}{Character name of the kernel.}
}
\value{
A list describing the kernel.
}
\description{
Get a registered kernel
}
