% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-kernel-registry.R
\name{get_kernel_def}
\alias{get_kernel_def}
\title{Get kernel definition}
\usage{
get_kernel_def(kernel)
}
\arguments{
\item{kernel}{Kernel name.}
}
\value{
A kernel definition list.
}
\description{
Get kernel definition
}
\keyword{internal}
