% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{globals}
\alias{globals}
\title{Internal global declarations}
\description{
Declares global variables to satisfy R CMD check when using non-standard
evaluation or generated column names.
}
\keyword{internal}
