% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-model-methods.R
\name{mcmc_ggdiag}
\alias{mcmc_ggdiag}
\title{Mcmc ggdiag}
\usage{
mcmc_ggdiag(object, what = c("trace", "density"), ...)
}
\arguments{
\item{object}{object.}

\item{what}{what.}

\item{...}{Additional arguments.}
}
\value{
See details.
}
\description{
Mcmc ggdiag.
}
\examples{
f <- getFromNamespace("mcmc_ggdiag", "DPmixGPD")
f
\dontrun{
# Example usage (may require MCMC run and nimble compilation).
}

}
