% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-S3-Methods.R
\name{plot.dpmixgpd_ate}
\alias{plot.dpmixgpd_ate}
\title{Plot ATE results}
\usage{
\method{plot}{dpmixgpd_ate}(x, y = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{dpmixgpd_ate}.}

\item{y}{Ignored.}

\item{...}{Additional arguments passed to ggplot2 functions.}
}
\value{
A list of ggplot objects.
}
\description{
Plot ATE results
}
