% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-S3-Methods.R
\name{plot.dpmixgpd_causal_fit}
\alias{plot.dpmixgpd_causal_fit}
\title{Plot a causal fit}
\usage{
\method{plot}{dpmixgpd_causal_fit}(x, arm = "both", ...)
}
\arguments{
\item{x}{A \code{"dpmixgpd_causal_fit"} object.}

\item{arm}{Integer or character; \code{1} or \code{"treated"} for treatment,
\code{0} or \code{"control"} for control.}

\item{...}{Additional arguments forwarded to the underlying outcome plot method.}
}
\value{
The result of the underlying plot call (invisibly).
}
\description{
Plot a causal fit
}
