% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_s3.R
\name{plot.dpmixgpd_cluster_bundle}
\alias{plot.dpmixgpd_cluster_bundle}
\title{Plot a cluster bundle}
\usage{
\method{plot}{dpmixgpd_cluster_bundle}(x, ...)
}
\arguments{
\item{x}{A cluster bundle.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Plot a cluster bundle
}
