% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-run-mcmc.R
\name{plot.dpmixgpd_fit}
\alias{plot.dpmixgpd_fit}
\title{Plot method for DPmixGPD fit (ggmcmc diagnostics)}
\usage{
\method{plot}{dpmixgpd_fit}(
  x,
  params = NULL,
  family = c("trace", "density", "running", "autocorr", "rhat", "ess"),
  ...
)
}
\arguments{
\item{x}{A \code{dpmixgpd_fit}.}

\item{params}{Optional character vector of parameters to plot (default: all).}

\item{family}{Optional subset for plot families: any of
c("trace","density","running","autocorr","rhat","ess")}

\item{...}{passed to ggmcmc geoms when applicable}
}
\description{
Plot method for DPmixGPD fit (ggmcmc diagnostics)
}
