% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-S3-Methods.R
\name{plot.dpmixgpd_qte}
\alias{plot.dpmixgpd_qte}
\title{Plot QTE results}
\usage{
\method{plot}{dpmixgpd_qte}(x, y = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{dpmixgpd_qte}.}

\item{y}{Ignored.}

\item{...}{Additional arguments passed to ggplot2 functions.}
}
\value{
A list of ggplot objects.
}
\description{
Plot QTE results
}
