% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-causal.R
\name{plot_te_curve}
\alias{plot_te_curve}
\title{Plot treatment-effect curve vs propensity score}
\usage{
plot_te_curve(df)
}
\arguments{
\item{df}{Data frame from te_curve().}
}
\value{
ggplot2 object.
}
\description{
Plot treatment-effect curve vs propensity score
}
