% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-causal.R
\name{predict.mixgpd_te_fit}
\alias{predict.mixgpd_te_fit}
\title{Predict mixgpd te fit}
\usage{
\method{predict}{mixgpd_te_fit}(
  object,
  newdata = NULL,
  type = c("quantile"),
  probs = c(0.1, 0.5, 0.9),
  alpha = object$alpha \%||\% 0.05,
  ...
)
}
\arguments{
\item{object}{object.}

\item{newdata}{newdata.}

\item{type}{type.}

\item{probs}{probs.}

\item{alpha}{alpha.}

\item{...}{Additional arguments.}
}
\value{
See details.
}
\description{
Predict mixgpd te fit.
}
\examples{
f <- getFromNamespace("predict.mixgpd_te_fit", "DPmixGPD")
f
\dontrun{
# Example usage (may require MCMC run and nimble compilation).
}

}
