% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-S3-Methods.R
\name{print.dpmixgpd_causal_fit}
\alias{print.dpmixgpd_causal_fit}
\title{Print a causal fit}
\usage{
\method{print}{dpmixgpd_causal_fit}(x, ...)
}
\arguments{
\item{x}{A \code{"dpmixgpd_causal_fit"} object.}

\item{...}{Unused.}
}
\value{
The input object (invisibly).
}
\description{
Print a causal fit
}
