% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-S3-Methods.R
\name{print.dpmixgpd_causal_predict_plots}
\alias{print.dpmixgpd_causal_predict_plots}
\title{Print method for causal prediction plots}
\usage{
\method{print}{dpmixgpd_causal_predict_plots}(x, ...)
}
\arguments{
\item{x}{Object of class \code{dpmixgpd_causal_predict_plots}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print method for causal prediction plots
}
\keyword{internal}
