% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_s3.R
\name{print.dpmixgpd_cluster_bundle}
\alias{print.dpmixgpd_cluster_bundle}
\title{Print a cluster bundle}
\usage{
\method{print}{dpmixgpd_cluster_bundle}(x, ...)
}
\arguments{
\item{x}{A cluster bundle.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a cluster bundle
}
