% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_s3.R
\name{print.dpmixgpd_cluster_labels}
\alias{print.dpmixgpd_cluster_labels}
\title{Print cluster labels}
\usage{
\method{print}{dpmixgpd_cluster_labels}(x, ...)
}
\arguments{
\item{x}{Cluster labels object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print cluster labels
}
