% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-run-mcmc.R
\name{print.dpmixgpd_fit}
\alias{print.dpmixgpd_fit}
\title{Print a DPmixGPD fit object}
\usage{
\method{print}{dpmixgpd_fit}(x, ...)
}
\arguments{
\item{x}{A \code{dpmixgpd_fit}.}

\item{...}{unused.}
}
\description{
Prints model description and prior table; does not print posterior samples.
}
