% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-S3-Methods.R
\name{print.dpmixgpd_ps_bundle}
\alias{print.dpmixgpd_ps_bundle}
\title{Print a propensity score bundle}
\usage{
\method{print}{dpmixgpd_ps_bundle}(x, code = FALSE, max_code_lines = 200L, ...)
}
\arguments{
\item{x}{A \code{"dpmixgpd_ps_bundle"} object.}

\item{code}{Logical; if TRUE, print generated NIMBLE code for the PS model.}

\item{max_code_lines}{Integer; maximum number of code lines to print when \code{code=TRUE}.}

\item{...}{Unused.}
}
\value{
The input object (invisibly).
}
\description{
Print a propensity score bundle
}
