% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-S3-Methods.R
\name{print.dpmixgpd_ps_fit}
\alias{print.dpmixgpd_ps_fit}
\title{Print a propensity score fit}
\usage{
\method{print}{dpmixgpd_ps_fit}(x, ...)
}
\arguments{
\item{x}{A \code{"dpmixgpd_ps_fit"} object.}

\item{...}{Unused.}
}
\value{
The input object (invisibly).
}
\description{
Print a propensity score fit
}
