% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.mixgpd_allocation_plots}
\alias{print.mixgpd_allocation_plots}
\title{Print cluster allocation plots}
\usage{
\method{print}{mixgpd_allocation_plots}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mixgpd_allocation_plots}.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\value{
Invisibly returns the input object.
}
\description{
Print cluster allocation plots
}
