% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-causal.R
\name{print.mixgpd_te}
\alias{print.mixgpd_te}
\title{Print mixgpd te}
\usage{
\method{print}{mixgpd_te}(x, ...)
}
\arguments{
\item{x}{x.}

\item{...}{Additional arguments.}
}
\value{
See details.
}
\description{
Print mixgpd te.
}
\examples{
f <- getFromNamespace("print.mixgpd_te", "DPmixGPD")
f
\dontrun{
# Example usage (may require MCMC run and nimble compilation).
}

}
