% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.causalmixgpd_causal_fit}
\alias{print.summary.causalmixgpd_causal_fit}
\title{Print a causal-model summary object}
\usage{
\method{print}{summary.causalmixgpd_causal_fit}(x, digits = 3, max_rows = 60, ...)
}
\arguments{
\item{x}{A \code{"summary.causalmixgpd_causal_fit"} object.}

\item{digits}{Number of digits to print in summary tables.}

\item{max_rows}{Maximum rows to print from each summary table.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print a causal-model summary object
}
