% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-S3-Methods.R
\name{print.summary.dpmixgpd_ate}
\alias{print.summary.dpmixgpd_ate}
\title{Print an ATE summary}
\usage{
\method{print}{summary.dpmixgpd_ate}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{"summary.dpmixgpd_ate"} object.}

\item{digits}{Number of digits to display.}

\item{...}{Unused.}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print an ATE summary
}
