% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-run-mcmc.R
\name{print.summary.dpmixgpd_fit}
\alias{print.summary.dpmixgpd_fit}
\title{Print summary.dpmixgpd_fit}
\usage{
\method{print}{summary.dpmixgpd_fit}(x, ...)
}
\arguments{
\item{x}{Summary object.}

\item{...}{unused.}
}
\description{
Print summary.dpmixgpd_fit
}
