% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-S3-Methods.R
\name{print.summary.dpmixgpd_qte}
\alias{print.summary.dpmixgpd_qte}
\title{Print a QTE summary}
\usage{
\method{print}{summary.dpmixgpd_qte}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{"summary.dpmixgpd_qte"} object.}

\item{digits}{Number of digits to display.}

\item{...}{Unused.}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print a QTE summary
}
