% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.mixgpd_allocation}
\alias{print.summary.mixgpd_allocation}
\title{Print summary of cluster allocation}
\usage{
\method{print}{summary.mixgpd_allocation}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.mixgpd_allocation}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print summary of cluster allocation
}
