% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-model-methods.R
\name{print.summary.mixgpd_fit}
\alias{print.summary.mixgpd_fit}
\title{Print summary mixgpd fit}
\usage{
\method{print}{summary.mixgpd_fit}(x, ...)
}
\arguments{
\item{x}{x.}

\item{...}{Additional arguments.}
}
\value{
See details.
}
\description{
Print summary mixgpd fit.
}
\examples{
f <- getFromNamespace("print.summary.mixgpd_fit", "DPmixGPD")
f
\dontrun{
# Example usage (may require MCMC run and nimble compilation).
}

}
