% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_effect.R
\name{qate}
\alias{qate}
\title{Quantile Average Treatment Effect (QATE)
(Alias of QTE in this package)}
\usage{
qate(object, ...)
}
\description{
Quantile Average Treatment Effect (QATE)
(Alias of QTE in this package)
}
