% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-engine.R
\name{run_mcmc_engine}
\alias{run_mcmc_engine}
\title{Run mcmc engine}
\usage{
run_mcmc_engine(spec, mcmc)
}
\arguments{
\item{spec}{spec.}

\item{mcmc}{mcmc.}
}
\value{
See details.
}
\description{
Run mcmc engine.
}
\examples{
f <- getFromNamespace("run_mcmc_engine", "DPmixGPD")
f

}
\keyword{internal}
