% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_gamma_nimble.R
\name{run_mcmc_nimble_gamma}
\alias{run_mcmc_nimble_gamma}
\title{Internal: run Nimble MCMC for Gamma kernel (uncond or regression)}
\usage{
run_mcmc_nimble_gamma(spec, mcmc)
}
\description{
Internal: run Nimble MCMC for Gamma kernel (uncond or regression)
}
\keyword{internal}
