% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_gamma_nimble.R
\name{run_mcmc_nimble_gamma_uncond}
\alias{run_mcmc_nimble_gamma_uncond}
\title{Internal: run Nimble MCMC for unconditional Gamma SB mixture}
\usage{
run_mcmc_nimble_gamma_uncond(spec, mcmc)
}
\description{
Internal: run Nimble MCMC for unconditional Gamma SB mixture
}
\keyword{internal}
