% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-simulated-data.R
\name{sim_causal_cqte}
\alias{sim_causal_cqte}
\title{Simulate causal CQTE data}
\usage{
sim_causal_cqte(n = 300, seed = NULL)
}
\arguments{
\item{n}{Sample size.}

\item{seed}{Optional RNG seed.}
}
\value{
A list with y, t, X.
}
\description{
Simulate causal CQTE data
}
