% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-engine.R
\name{simulate_chain}
\alias{simulate_chain}
\title{Simulate chain}
\usage{
simulate_chain(spec, n_iter, burn_in)
}
\arguments{
\item{spec}{spec.}

\item{n_iter}{n_iter.}

\item{burn_in}{burn_in.}
}
\value{
See details.
}
\description{
Simulate chain.
}
\examples{
f <- getFromNamespace("simulate_chain", "DPmixGPD")
f

}
\keyword{internal}
