% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_s3.R
\name{summary.dpmixgpd_cluster_bundle}
\alias{summary.dpmixgpd_cluster_bundle}
\title{Summarize a cluster bundle}
\usage{
\method{summary}{dpmixgpd_cluster_bundle}(object, ...)
}
\arguments{
\item{object}{A cluster bundle.}

\item{...}{Unused.}
}
\value{
A summary list.
}
\description{
Summarize a cluster bundle
}
