% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_s3.R
\name{summary.dpmixgpd_cluster_labels}
\alias{summary.dpmixgpd_cluster_labels}
\title{Summarize cluster labels}
\usage{
\method{summary}{dpmixgpd_cluster_labels}(object, ...)
}
\arguments{
\item{object}{Cluster labels object.}

\item{...}{Unused.}
}
\value{
A summary list.
}
\description{
Summarize cluster labels
}
