% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_s3.R
\name{summary.dpmixgpd_cluster_psm}
\alias{summary.dpmixgpd_cluster_psm}
\title{Summarize a cluster posterior similarity matrix}
\usage{
\method{summary}{dpmixgpd_cluster_psm}(object, ...)
}
\arguments{
\item{object}{Cluster PSM object.}

\item{...}{Unused.}
}
\value{
A summary list.
}
\description{
Summarize a cluster posterior similarity matrix
}
