% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-run-mcmc.R
\name{summary.dpmixgpd_fit}
\alias{summary.dpmixgpd_fit}
\title{Summary of DPmixGPD fit}
\usage{
\method{summary}{dpmixgpd_fit}(object, ...)
}
\arguments{
\item{object}{A \code{dpmixgpd_fit}.}

\item{...}{unused.}
}
\description{
Prints model description, parameter summary (from MCMC samples), and WAIC if available.
}
