% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.mixgpd_allocation}
\alias{summary.mixgpd_allocation}
\title{Summarize cluster allocation}
\usage{
\method{summary}{mixgpd_allocation}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mixgpd_allocation}.}

\item{...}{Additional arguments (ignored).}
}
\value{
An object of class \code{summary.mixgpd_allocation} containing detailed
summaries of cluster labels and probabilities.
}
\description{
Summarize cluster allocation
}
