% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R, R/treatment_effect.R
\name{ate}
\alias{ate}
\title{Average Treatment Effect (ATE) for a fitted TE object}
\usage{
ate(object, ...)

ate(object, ...)
}
\description{
Average Treatment Effect (ATE) for a fitted TE object

Average Treatment Effect (ATE)
}
