% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-engine-sb-gamma.R
\name{build_nimble_model_gamma_reg}
\alias{build_nimble_model_gamma_reg}
\title{Build nimble model gamma reg}
\usage{
build_nimble_model_gamma_reg(spec)
}
\arguments{
\item{spec}{spec.}
}
\value{
See details.
}
\description{
Build nimble model gamma reg.
}
\examples{
f <- getFromNamespace("build_nimble_model_gamma_reg", "DPmixGPD")
f

}
\keyword{internal}
