% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-engine.R
\name{build_nimble_model_gamma_uncond}
\alias{build_nimble_model_gamma_uncond}
\title{Build nimble model gamma uncond}
\usage{
build_nimble_model_gamma_uncond(spec)
}
\arguments{
\item{spec}{spec.}
}
\value{
See details.
}
\description{
Build nimble model gamma uncond.
}
\examples{
f <- getFromNamespace("build_nimble_model_gamma_uncond", "DPmixGPD")
f

}
\keyword{internal}
