% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{.check_formula_data}
\alias{.check_formula_data}
\title{Check formula data}
\usage{
.check_formula_data(formula, data)
}
\arguments{
\item{formula}{formula.}

\item{data}{data.}
}
\value{
See details.
}
\description{
Check formula data.
}
\examples{
f <- getFromNamespace(".check_formula_data", "DPmixGPD")
f

}
\keyword{internal}
