% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.check_formula_data}
\alias{.check_formula_data}
\title{Internal checks: formula/data consistency}
\usage{
.check_formula_data(formula, data)
}
\description{
Internal checks: formula/data consistency
}
\keyword{internal}
