% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.check_kernel_support}
\alias{.check_kernel_support}
\title{Internal checks: kernel vs response support}
\usage{
.check_kernel_support(Y, kernel)
}
\description{
Internal checks: kernel vs response support
}
\keyword{internal}
