% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncore-utils.R
\name{.check_mcmc}
\alias{.check_mcmc}
\title{Check MCMC control list}
\usage{
.check_mcmc(mcmc)
}
\arguments{
\item{mcmc}{A list with at least \code{niter} and \code{nburn}.}
}
\value{
A normalized list.
}
\description{
Check MCMC control list
}
\keyword{internal}
