% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.detect_first_present}
\alias{.detect_first_present}
\title{Detect the first present argument name in dots.}
\usage{
.detect_first_present(dots, candidates = c("q", "x"))
}
\description{
Detect the first present argument name in dots.
}
\keyword{internal}
