% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-utilities-internal.R
\name{.extract_draws_matrix}
\alias{.extract_draws_matrix}
\title{Extract posterior draws as a numeric matrix (iter x parameters)}
\usage{
.extract_draws_matrix(object, drop_v = TRUE)
}
\arguments{
\item{object}{A mixgpd_fit.}

\item{drop_v}{Logical; if TRUE, drop stick-breaking v parameters.}
}
\value{
Numeric matrix of draws.
}
\description{
Extract posterior draws as a numeric matrix (iter x parameters)
}
\keyword{internal}
