% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-utilities-internal.R
\name{.extract_nimble_code}
\alias{.extract_nimble_code}
\title{Extract nimbleCode from bundle code}
\usage{
.extract_nimble_code(code)
}
\description{
Extract nimbleCode from bundle code
}
\keyword{internal}
