% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-utilities-internal.R
\name{.format_fit_header}
\alias{.format_fit_header}
\title{Format a short header for printing}
\usage{
.format_fit_header(x)
}
\arguments{
\item{x}{A mixgpd_fit.}
}
\value{
Character vector lines.
}
\description{
Format a short header for printing
}
\keyword{internal}
