% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-utilities-internal.R
\name{.get_epsilon}
\alias{.get_epsilon}
\title{Get epsilon value from object spec/meta or argument}
\usage{
.get_epsilon(object, epsilon = NULL)
}
\arguments{
\item{object}{A mixgpd_fit object.}

\item{epsilon}{Numeric; if provided, overrides object spec/meta.}
}
\description{
Get epsilon value from object spec/meta or argument
}
\keyword{internal}
