% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-utilities-internal.R
\name{.get_nobs}
\alias{.get_nobs}
\title{Get number of observations used in fitting}
\usage{
.get_nobs(object)
}
\arguments{
\item{object}{A mixgpd_fit.}
}
\value{
Integer n.
}
\description{
Get number of observations used in fitting
}
\keyword{internal}
