% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.get_samples_mcmclist}
\alias{.get_samples_mcmclist}
\title{Safely coerce MCMC samples to coda::mcmc.list}
\usage{
.get_samples_mcmclist(object)
}
\arguments{
\item{object}{A mixgpd_fit.}
}
\value{
A coda::mcmc.list object.
}
\description{
Safely coerce MCMC samples to coda::mcmc.list
}
\keyword{internal}
