% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{OnLoad}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Character; library path.}

\item{pkgname}{Character; package name.}
}
\value{
See details.

Invisibly returns `NULL`.
}
\description{
OnLoad.
}
\examples{
f <- getFromNamespace(".onLoad", "DPmixGPD")
f

getFromNamespace(".onLoad", "DPmixGPD")

}
\keyword{CMD}
\keyword{Initializes}
\keyword{On}
\keyword{`R}
\keyword{and}
\keyword{by}
\keyword{check`.}
\keyword{code}
\keyword{declares}
\keyword{generated}
\keyword{hook}
\keyword{internal}
\keyword{load}
\keyword{registries}
\keyword{satisfy}
\keyword{to}
\keyword{used}
\keyword{variables}
