% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Initializes internal registries and declares variables used by generated code
to satisfy `R CMD check`.}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Character; library path.}

\item{pkgname}{Character; package name.}
}
\value{
Invisibly returns `NULL`.
}
\description{
Initializes internal registries and declares variables used by generated code
to satisfy `R CMD check`.
}
\examples{
getFromNamespace(".onLoad", "DPmixGPD")

}
\keyword{internal}
