% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{.prepare_design_matrix}
\alias{.prepare_design_matrix}
\title{Prepare design matrix}
\usage{
.prepare_design_matrix(X, intercept = TRUE)
}
\arguments{
\item{X}{X.}

\item{intercept}{intercept.}
}
\value{
See details.
}
\description{
Prepare design matrix.
}
\examples{
f <- getFromNamespace(".prepare_design_matrix", "DPmixGPD")
f

}
\keyword{internal}
