% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-quantile-gamma-dp.R
\name{.q_gamma_dp_one_draw}
\alias{.q_gamma_dp_one_draw}
\title{Quantile of a Gamma DP mixture for ONE posterior draw}
\usage{
.q_gamma_dp_one_draw(
  p,
  w,
  shape,
  scale,
  lower = 0,
  upper0 = 1,
  expand = 2,
  max_expand = 60,
  tol = .Machine$double.eps^0.5,
  maxiter = 1000
)
}
\description{
Quantile of a Gamma DP mixture for ONE posterior draw
}
\keyword{internal}
