% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelspec-dp.R
\name{.q_gamma_dp_over_draws}
\alias{.q_gamma_dp_over_draws}
\title{Q gamma dp over draws}
\usage{
.q_gamma_dp_over_draws(
  p_vec,
  W,
  Shape,
  Scale,
  lower = 0,
  upper0 = 1,
  expand = 2,
  max_expand = 60,
  tol = .Machine$double.eps^0.5,
  maxiter = 1000
)
}
\arguments{
\item{p_vec}{p_vec.}

\item{W}{W.}

\item{Shape}{Shape.}

\item{Scale}{Scale.}

\item{lower}{lower.}

\item{upper0}{upper0.}

\item{expand}{expand.}

\item{max_expand}{max_expand.}

\item{tol}{tol.}

\item{maxiter}{maxiter.}
}
\value{
See details.
}
\description{
Q gamma dp over draws.
}
\examples{
f <- getFromNamespace(".q_gamma_dp_over_draws", "DPmixGPD")
f

}
\keyword{internal}
