% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_engine.R
\name{.transform_generate_code}
\alias{.transform_generate_code}
\title{Generate nimble code implementing a transform}
\usage{
.transform_generate_code(lhs, lp, transform)
}
\arguments{
\item{lhs}{Character, left-hand side (e.g. 'sigma[i]', 'shape[i,j]').}

\item{lp}{Character, linear predictor (e.g. 'lp_sigma[i]').}

\item{transform}{Character, one of .transform_valid_names().}
}
\value{
A character string containing an R/NIMBLE expression,
  e.g. 'sigma[i] <- exp(lp_sigma[i])'.
}
\description{
Generate nimble code implementing a transform
}
