% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_engine.R
\name{.transform_resolve}
\alias{.transform_resolve}
\title{Resolve transform for a given parameter}
\usage{
.transform_resolve(X, trans, caller = "fit.dpm")
}
\arguments{
\item{trans}{List supplied by user to fit.dpm()/fit.mixgpd(), e.g.
list(scale = "exp", shape = "identity").}

\item{param_name}{Character, name of parameter ("scale", "shape",
"threshold", "xi", "sigma", etc.).}

\item{default}{Character, default transform if none is provided.}

\item{must_be_positive}{Logical, whether this parameter is required
to be positive for the chosen kernel/tail.}
}
\value{
A single canonical transform name.
}
\description{
Resolve transform for a given parameter
}
