% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.truncate_draws_matrix_components}
\alias{.truncate_draws_matrix_components}
\title{Truncate component draws in a draws matrix}
\usage{
.truncate_draws_matrix_components(object, mat, epsilon)
}
\arguments{
\item{object}{A mixgpd_fit object.}

\item{mat}{Numeric matrix of draws (iter x parameters).}

\item{epsilon}{Numeric in [0,1). Truncation level.}
}
\value{
Numeric matrix with truncated components.
}
\description{
Truncate component draws in a draws matrix
}
\keyword{internal}
