% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.validate_fit}
\alias{.validate_fit}
\title{Validate a fitted object}
\usage{
.validate_fit(object)
}
\arguments{
\item{object}{A fitted object.}
}
\value{
Invisibly TRUE, otherwise errors.
}
\description{
Validate a fitted object
}
\keyword{internal}
