% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.wrap_scalar_first_arg}
\alias{.wrap_scalar_first_arg}
\title{Wrap scalar first-argument functions to handle vector inputs.}
\usage{
.wrap_scalar_first_arg(fun, first_arg_name)
}
\description{
Wrap scalar first-argument functions to handle vector inputs.
}
\keyword{internal}
