% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-utilities-internal.R
\name{.wrap_scalar_p}
\alias{.wrap_scalar_p}
\title{Wrap scalar CDF to handle q/x naming and vector inputs.}
\usage{
.wrap_scalar_p(fun)
}
\description{
Wrap scalar CDF to handle q/x naming and vector inputs.
}
\keyword{internal}
