% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPmixGPD-vignettes.R
\name{DPmixGPD-vignettes}
\alias{DPmixGPD-vignettes}
\alias{basic-vignette}
\alias{unconditional-vignette}
\alias{conditional-vignette}
\alias{causal-vignette}
\title{Vignettes for DPmixGPD}
\description{
Entry points to the package vignettes.
}
\details{
You can list all vignettes with \code{browseVignettes("DPmixGPD")}.

Open from R:
\itemize{
\item \code{vignette("basic-01", package = "DPmixGPD")}
\item \code{vignette("unconditional-03", package = "DPmixGPD")}
\item \code{vignette("conditional-04", package = "DPmixGPD")}
\item \code{vignette("causal-05", package = "DPmixGPD")}
}

Website articles index:
\itemize{
\item \url{https://arnabaich96.github.io/DPmixGPD/pkgdown/articles/index.html}
}
}
