% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPmixGPD-vignettes.R
\name{DPmixGPD-vignettes}
\alias{DPmixGPD-vignettes}
\alias{basic-vignette}
\alias{unconditional-vignette}
\alias{conditional-vignette}
\alias{causal-vignette}
\title{Vignettes for DPmixGPD}
\description{
Entry points to the package vignettes.
}
\details{
You can list all vignettes with \code{browseVignettes("DPmixGPD")}.

Open from R:
\itemize{
\item \code{vignette("01-basic", package = "DPmixGPD")}
\item \code{vignette("03-unconditional", package = "DPmixGPD")}
\item \code{vignette("04-conditional", package = "DPmixGPD")}
\item \code{vignette("05-causal", package = "DPmixGPD")}
}

Installed vignette pages:
\itemize{
\item \ifelse{html}{\href{../doc/01-basic.html}{01-basic.html}}{\code{01-basic.html}}
\item \ifelse{html}{\href{../doc/03-unconditional.html}{03-unconditional.html}}{\code{03-unconditional.html}}
\item \ifelse{html}{\href{../doc/04-conditional.html}{04-conditional.html}}{\code{04-conditional.html}}
\item \ifelse{html}{\href{../doc/05-causal.html}{05-causal.html}}{\code{05-causal.html}}
}
}
