% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPmixGPD-package.R
\docType{package}
\name{DPmixGPD}
\alias{DPmixGPD}
\alias{DPmixGPD-package}
\title{DPmixGPD: Dirichlet Process Mixtures with Generalized Pareto Tail}
\description{
\code{DPmixGPD} provides Bayesian Dirichlet process mixture models for
unconditional, conditional, and causal workflows, with optional generalized
Pareto tail modeling.
}
\details{
Core entry points include \code{\link[=build_nimble_bundle]{build_nimble_bundle()}}, \code{\link[=run_mcmc_bundle_manual]{run_mcmc_bundle_manual()}},
\code{\link[=build_causal_bundle]{build_causal_bundle()}}, and \code{\link[=run_mcmc_causal]{run_mcmc_causal()}}.

Learn more via:
\itemize{
\item package index: \code{help(package = "DPmixGPD")}
\item vignettes: \code{browseVignettes("DPmixGPD")}
\item website articles index: \url{https://arnabaich96.github.io/DPmixGPD/pkgdown/articles/index.html}
\item vignette pages:
\itemize{
\item \code{vignette("basic-01", package = "DPmixGPD")}
\item \code{vignette("unconditional-03", package = "DPmixGPD")}
\item \code{vignette("conditional-04", package = "DPmixGPD")}
\item \code{vignette("causal-05", package = "DPmixGPD")}
}
\item website: \url{https://arnabaich96.github.io/DPmixGPD/}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://arnabaich96.github.io/DPmixGPD/}
  \item \url{https://arnabaich96.github.io/DPmixGPD/pkgdown/}
  \item Report bugs at \url{https://github.com/arnabaich96/DPmixGPD/issues}
}

}
\author{
\strong{Maintainer}: Arnab Aich \email{aaich@fsu.edu}

}
