% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPmixGPD-package.R
\docType{package}
\name{DPmixGPD}
\alias{DPmixGPD}
\alias{DPmixGPD-package}
\title{DPmixGPD: Dirichlet Process Mixtures with Generalized Pareto Tail}
\description{
\code{DPmixGPD} provides Bayesian Dirichlet process mixture models for
unconditional, conditional, and causal models, with optional generalized
Pareto tail modeling.
Learn more via:
\itemize{
\item package index: \code{help(package = "DPmixGPD")}
\item vignettes: \code{browseVignettes("DPmixGPD")}
\item pkgdown_site: \url{https://arnabaich96.github.io/DPmixGPD/pkgdown/articles/index.html}
\item vignettes:
\itemize{
\item \href{../doc/basic.html}{Model specification and posterior computation}
\item \href{../doc/model-spec.html}{Spliced Model fundamentals}
\item \href{../doc/unconditional.html}{Unconditional model}
\item \href{../doc/conditional.html}{Conditional model}
\item \href{../doc/causal.html}{Causal model}
}
\item website: \url{https://arnabaich96.github.io/DPmixGPD/}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://arnabaich96.github.io/DPmixGPD/}
  \item \url{https://arnabaich96.github.io/DPmixGPD/pkgdown/}
  \item Report bugs at \url{https://github.com/arnabaich96/DPmixGPD/issues}
}

}
\author{
\strong{Maintainer}: Arnab Aich \email{aaich@fsu.edu}

}
