% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dpm_wrappers.R
\name{fit.dpmgpd}
\alias{fit.dpmgpd}
\title{Dirichlet process mixture with GPD tail}
\usage{
fit.dpmgpd(
  formula,
  data,
  kernel = c("gamma", "lognormal", "normal", "laplace", "invgauss", "amoroso"),
  dp_rep = c("stick_breaking", "crp"),
  dp_ctrl = list(),
  priors = list(),
  tail_ctrl = list(),
  tail_priors = list(),
  trans = list(),
  mcmc = list(),
  alpha = 0.05
)
}
\description{
Fit a Dirichlet process mixture for the bulk with a GPD tail for exceedances.
}
