% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_functions.R
\name{fit_mixgpd}
\alias{fit_mixgpd}
\title{Fit DP mixture model without GPD tail (formula interface)}
\usage{
fit_mixgpd(
  formula,
  data,
  kernel = c("gamma", "lognormal", "normal", "laplace", "inverse_gaussian", "amoroso",
    "pareto"),
  dp_rep = c("stick_breaking", "crp"),
  dp_ctrl = list(),
  priors = list(),
  trans = list(),
  mcmc = list(),
  alpha = 0.05,
  seed = NULL
)
}
\description{
Fit DP mixture model without GPD tail (formula interface)
}
