% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-model-methods.R
\name{fit_mixgpd_gpd}
\alias{fit_mixgpd_gpd}
\title{Fit mixgpd gpd}
\usage{
fit_mixgpd_gpd(
  formula,
  data,
  kernel = c("gamma", "lognormal", "normal", "laplace", "inverse_gaussian", "amoroso",
    "pareto"),
  dp_rep = c("stick_breaking", "crp"),
  dp_ctrl = list(),
  priors = list(),
  trans = list(),
  mcmc = list(),
  alpha = 0.05,
  seed = NULL
)
}
\arguments{
\item{formula}{formula.}

\item{data}{data.}

\item{kernel}{kernel.}

\item{dp_rep}{dp_rep.}

\item{dp_ctrl}{dp_ctrl.}

\item{priors}{priors.}

\item{trans}{trans.}

\item{mcmc}{mcmc.}

\item{alpha}{alpha.}

\item{seed}{seed.}
}
\value{
See details.
}
\description{
Fit mixgpd gpd.
}
\examples{
f <- getFromNamespace("fit_mixgpd_gpd", "DPmixGPD")
f
\dontrun{
# Example usage (may require MCMC run and nimble compilation).
}

}
