% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-utilities-internal.R, R/04-S3-Methods.R
\name{fitted.mixgpd_fit}
\alias{fitted.mixgpd_fit}
\title{Fitted values for a MixGPD fit}
\usage{
\method{fitted}{mixgpd_fit}(object, ...)

\method{fitted}{mixgpd_fit}(object, ...)
}
\arguments{
\item{object}{A fitted object of class \code{"mixgpd_fit"}.}

\item{...}{Unused.}
}
\value{
Numeric vector of length \code{nobs(object)}.

Numeric vector of length \code{nobs(object)}.
}
\description{
Returns a simple fitted value summary on the training data. If an engine
provides a fitted function at \code{object$engine$fitted}, it will be used.
Otherwise, returns the empirical median of \code{y} replicated.

Returns a simple fitted value summary on the training data. If an engine
provides a fitted function at \code{object$engine$fitted}, it will be used.
Otherwise, returns the empirical median of \code{y} replicated.
}
