% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-model-methods.R
\name{fitted.mixgpd_fit}
\alias{fitted.mixgpd_fit}
\title{Fitted mixgpd fit}
\usage{
\method{fitted}{mixgpd_fit}(object, ...)
}
\arguments{
\item{object}{object.}

\item{...}{Additional arguments.}
}
\value{
See details.
}
\description{
Fitted mixgpd fit.
}
\examples{
f <- getFromNamespace("fitted.mixgpd_fit", "DPmixGPD")
f
\dontrun{
# Example usage (may require MCMC run and nimble compilation).
}

}
