% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-kernel-registry.R
\name{get_kernel}
\alias{get_kernel}
\title{Get kernel}
\usage{
get_kernel(kernel)
}
\arguments{
\item{kernel}{kernel.}
}
\value{
See details.
}
\description{
Get kernel.
}
\examples{
f <- getFromNamespace("get_kernel", "DPmixGPD")
f
\dontrun{
# Example usage (may require MCMC run and nimble compilation).
}

}
