% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-kernel-registry.R
\name{get_kernel_registry}
\alias{get_kernel_registry}
\title{Get kernel registry}
\usage{
get_kernel_registry()
}
\value{
A list of kernel metadata.
}
\description{
Get kernel registry
}
\examples{
init_kernel_registry()
reg <- get_kernel_registry()
reg$normal$bulk_params
}
