% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-registry.R
\name{get_tail_registry}
\alias{get_tail_registry}
\title{Get tail registry}
\usage{
get_tail_registry()
}
\value{
A list of tail metadata.
}
\description{
Get tail registry
}
\examples{
init_kernel_registry()
tail <- get_tail_registry()
tail$params
}
