% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-kernel-registry.R
\name{init_kernel_registry}
\alias{init_kernel_registry}
\title{Init kernel registry}
\usage{
init_kernel_registry()
}
\value{
See details.
}
\description{
Init kernel registry.
}
\examples{
f <- getFromNamespace("init_kernel_registry", "DPmixGPD")
f

}
\keyword{internal}
