% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mcmc_ggdiag}
\alias{mcmc_ggdiag}
\title{Basic MCMC diagnostics using coda}
\usage{
mcmc_ggdiag(object, what = c("trace", "density"), ...)
}
\arguments{
\item{object}{mixgpd_fit}

\item{what}{character, "trace" or "density"}
}
\description{
Basic MCMC diagnostics using coda
}
