% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.causalmixgpd_causal_predict}
\alias{plot.causalmixgpd_causal_predict}
\title{Plot causal prediction outputs}
\usage{
\method{plot}{causalmixgpd_causal_predict}(x, y = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{causalmixgpd_causal_predict}.}

\item{y}{Ignored.}

\item{...}{Additional arguments passed to ggplot2 functions.}
}
\value{
A ggplot object or a list of ggplot objects.
}
\description{
S3 method for visualizing causal predictions from \code{predict.causalmixgpd_causal_fit()}.
For mean/quantile, plots treated/control and treatment effect versus PS (or index).
For density/prob, plots treated/control values versus y.
}
