% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{plot.dpmixgpd_cluster_bundle}
\alias{plot.dpmixgpd_cluster_bundle}
\title{Plot a cluster bundle}
\usage{
\method{plot}{dpmixgpd_cluster_bundle}(x, ...)
}
\arguments{
\item{x}{A cluster bundle.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Produce a compact graphical summary of the cluster bundle metadata.
}
\seealso{
\code{\link[=summary.dpmixgpd_cluster_bundle]{summary.dpmixgpd_cluster_bundle()}}, \code{\link[=dpmix.cluster]{dpmix.cluster()}}, \code{\link[=dpmgpd.cluster]{dpmgpd.cluster()}}.

Other cluster workflow: 
\code{\link{dpmgpd.cluster}()},
\code{\link{dpmix.cluster}()},
\code{\link{plot.dpmixgpd_cluster_fit}()},
\code{\link{plot.dpmixgpd_cluster_labels}()},
\code{\link{plot.dpmixgpd_cluster_psm}()},
\code{\link{predict.dpmixgpd_cluster_fit}()},
\code{\link{print.dpmixgpd_cluster_bundle}()},
\code{\link{print.dpmixgpd_cluster_fit}()},
\code{\link{print.dpmixgpd_cluster_labels}()},
\code{\link{print.dpmixgpd_cluster_psm}()},
\code{\link{summary.dpmixgpd_cluster_bundle}()},
\code{\link{summary.dpmixgpd_cluster_fit}()},
\code{\link{summary.dpmixgpd_cluster_labels}()},
\code{\link{summary.dpmixgpd_cluster_psm}()}
}
\concept{cluster workflow}
