% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_s3.R
\name{plot.dpmixgpd_cluster_labels}
\alias{plot.dpmixgpd_cluster_labels}
\title{Plot cluster labels}
\usage{
\method{plot}{dpmixgpd_cluster_labels}(x, type = c("sizes", "certainty"), ...)
}
\arguments{
\item{x}{Cluster labels object.}

\item{type}{Plot type.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Plot cluster labels
}
