% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_s3.R
\name{plot.dpmixgpd_cluster_labels}
\alias{plot.dpmixgpd_cluster_labels}
\title{Plot cluster labels}
\usage{
\method{plot}{dpmixgpd_cluster_labels}(x, type = c("sizes", "certainty"), ...)
}
\arguments{
\item{x}{Cluster labels object.}

\item{type}{Plot type.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Visualize representative cluster sizes or assignment certainty.
}
\seealso{
\code{\link[=summary.dpmixgpd_cluster_labels]{summary.dpmixgpd_cluster_labels()}}, \code{\link[=predict.dpmixgpd_cluster_fit]{predict.dpmixgpd_cluster_fit()}}.

Other cluster workflow: 
\code{\link{dpmgpd.cluster}()},
\code{\link{dpmix.cluster}()},
\code{\link{plot.dpmixgpd_cluster_bundle}()},
\code{\link{plot.dpmixgpd_cluster_fit}()},
\code{\link{plot.dpmixgpd_cluster_psm}()},
\code{\link{predict.dpmixgpd_cluster_fit}()},
\code{\link{print.dpmixgpd_cluster_bundle}()},
\code{\link{print.dpmixgpd_cluster_fit}()},
\code{\link{print.dpmixgpd_cluster_labels}()},
\code{\link{print.dpmixgpd_cluster_psm}()},
\code{\link{summary.dpmixgpd_cluster_bundle}()},
\code{\link{summary.dpmixgpd_cluster_fit}()},
\code{\link{summary.dpmixgpd_cluster_labels}()},
\code{\link{summary.dpmixgpd_cluster_psm}()}
}
\concept{cluster workflow}
