% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_s3.R
\name{plot.dpmixgpd_cluster_psm}
\alias{plot.dpmixgpd_cluster_psm}
\title{Plot a cluster posterior similarity matrix}
\usage{
\method{plot}{dpmixgpd_cluster_psm}(x, psm_max_n = x$psm_max_n \%||\% 2000L, ...)
}
\arguments{
\item{x}{Cluster PSM object.}

\item{psm_max_n}{Maximum allowed matrix size for plotting.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Plot a cluster posterior similarity matrix
}
