% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-plot-fit.R
\name{plot.mixgpd_fit}
\alias{plot.mixgpd_fit}
\title{Plot method for mixgpd_fit objects}
\usage{
\method{plot}{mixgpd_fit}(x, type = c("density"), grid_n = 200, ...)
}
\arguments{
\item{x}{A `mixgpd_fit` object.}

\item{type}{Plot type. Currently supports `"density"`.}

\item{grid_n}{Number of grid points for density evaluation.}

\item{...}{Unused.}
}
\value{
A `ggplot` object.
}
\description{
Currently supports predictive density plots for response-only Gamma DPM fits.
}
