% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-model-methods.R
\name{predict.mixgpd_fit}
\alias{predict.mixgpd_fit}
\title{Predict mixgpd fit}
\usage{
\method{predict}{mixgpd_fit}(
  object,
  newdata = NULL,
  type = c("quantile", "density", "cdf", "sample"),
  probs = c(0.1, 0.5, 0.9),
  n_samples = 1000L,
  level = NULL,
  renormalize_weights = TRUE,
  ...
)
}
\arguments{
\item{object}{object.}

\item{newdata}{newdata.}

\item{type}{type.}

\item{probs}{probs.}

\item{n_samples}{n_samples.}

\item{level}{level of significance}

\item{renormalize_weights}{renormalize_weights(default: TRUE).}

\item{...}{Additional arguments.}
}
\value{
See details.
}
\description{
Predict mixgpd fit.
}
\examples{
f <- getFromNamespace("predict.mixgpd_fit", "DPmixGPD")
f
\dontrun{
# Example usage (may require MCMC run and nimble compilation).
}

}
