% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.causalmixgpd_causal_fit}
\alias{print.causalmixgpd_causal_fit}
\title{Print a causal fit}
\usage{
\method{print}{causalmixgpd_causal_fit}(x, ...)
}
\arguments{
\item{x}{A \code{"causalmixgpd_causal_fit"} object.}

\item{...}{Unused.}
}
\value{
The input object (invisibly).
}
\description{
Print a causal fit
}
