% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_s3.R
\name{print.dpmixgpd_cluster_fit}
\alias{print.dpmixgpd_cluster_fit}
\title{Print a cluster fit}
\usage{
\method{print}{dpmixgpd_cluster_fit}(x, ...)
}
\arguments{
\item{x}{A cluster fit.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a cluster fit
}
