% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{print.dpmixgpd_cluster_labels}
\alias{print.dpmixgpd_cluster_labels}
\title{Print cluster labels}
\usage{
\method{print}{dpmixgpd_cluster_labels}(x, ...)
}
\arguments{
\item{x}{Cluster labels object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Compact display for a representative clustering.
}
\seealso{
\code{\link[=predict.dpmixgpd_cluster_fit]{predict.dpmixgpd_cluster_fit()}}, \code{\link[=summary.dpmixgpd_cluster_labels]{summary.dpmixgpd_cluster_labels()}},
\code{\link[=plot.dpmixgpd_cluster_labels]{plot.dpmixgpd_cluster_labels()}}.

Other cluster workflow: 
\code{\link{dpmgpd.cluster}()},
\code{\link{dpmix.cluster}()},
\code{\link{plot.dpmixgpd_cluster_bundle}()},
\code{\link{plot.dpmixgpd_cluster_fit}()},
\code{\link{plot.dpmixgpd_cluster_labels}()},
\code{\link{plot.dpmixgpd_cluster_psm}()},
\code{\link{predict.dpmixgpd_cluster_fit}()},
\code{\link{print.dpmixgpd_cluster_bundle}()},
\code{\link{print.dpmixgpd_cluster_fit}()},
\code{\link{print.dpmixgpd_cluster_psm}()},
\code{\link{summary.dpmixgpd_cluster_bundle}()},
\code{\link{summary.dpmixgpd_cluster_fit}()},
\code{\link{summary.dpmixgpd_cluster_labels}()},
\code{\link{summary.dpmixgpd_cluster_psm}()}
}
\concept{cluster workflow}
