% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_s3.R
\name{print.dpmixgpd_cluster_psm}
\alias{print.dpmixgpd_cluster_psm}
\title{Print a cluster posterior similarity matrix}
\usage{
\method{print}{dpmixgpd_cluster_psm}(x, ...)
}
\arguments{
\item{x}{Cluster PSM object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a cluster posterior similarity matrix
}
