% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.mixgpd_fit}
\alias{print.mixgpd_fit}
\title{Print a MixGPD fitted object}
\usage{
\method{print}{mixgpd_fit}(x, ...)
}
\arguments{
\item{x}{A fitted object of class \code{"mixgpd_fit"}.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print a MixGPD fitted object
}
\examples{
\dontrun{
y <- abs(stats::rnorm(50)) + 0.1
bundle <- build_nimble_bundle(y = y, backend = "sb", kernel = "normal",
                             GPD = TRUE, components = 6,
                             mcmc = list(niter = 200, nburnin = 50, thin = 1, nchains = 1))
fit <- run_mcmc_bundle_manual(bundle)
print(fit)
}
}
