% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-model-methods.R
\name{print.mixgpd_fit}
\alias{print.mixgpd_fit}
\title{Print summary of a MixGPD fit}
\usage{
\method{print}{mixgpd_fit}(x, ...)
}
\arguments{
\item{x}{An object of class `summary.mixgpd_fit`.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns `x`.
}
\description{
Print summary of a MixGPD fit
}
