% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-S3-Methods.R
\name{print.mixgpd_fit}
\alias{print.mixgpd_fit}
\title{Print a MixGPD fitted object}
\usage{
\method{print}{mixgpd_fit}(x, ...)
}
\arguments{
\item{x}{A fitted object of class \code{"mixgpd_fit"}.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print a MixGPD fitted object
}
