% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-S3-Methods.R
\name{print.mixgpd_fit_plots}
\alias{print.mixgpd_fit_plots}
\title{Print method for mixgpd_fit diagnostic plots}
\usage{
\method{print}{mixgpd_fit_plots}(x, ...)
}
\arguments{
\item{x}{Object of class \code{mixgpd_fit_plots}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print method for mixgpd_fit diagnostic plots
}
\keyword{internal}
