% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-S3-Methods.R
\name{print.mixgpd_fitted_plots}
\alias{print.mixgpd_fitted_plots}
\title{Print method for fitted value plots}
\usage{
\method{print}{mixgpd_fitted_plots}(x, ...)
}
\arguments{
\item{x}{Object of class \code{mixgpd_fitted_plots}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print method for fitted value plots
}
