% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.mixgpd_predict_plots}
\alias{print.mixgpd_predict_plots}
\title{Print method for prediction plots}
\usage{
\method{print}{mixgpd_predict_plots}(x, ...)
}
\arguments{
\item{x}{Object of class \code{mixgpd_predict_plots}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print method for prediction plots
}
