% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-S3-Methods.R
\name{print.mixgpd_summary}
\alias{print.mixgpd_summary}
\title{Print a MixGPD summary object}
\usage{
\method{print}{mixgpd_summary}(x, digits = 3, max_rows = 60, ...)
}
\arguments{
\item{x}{A \code{"mixgpd_summary"} object.}

\item{digits}{Number of digits to print.}

\item{max_rows}{Maximum rows to print.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print a MixGPD summary object
}
