% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.causalmixgpd_ps_fit}
\alias{print.summary.causalmixgpd_ps_fit}
\title{Print a propensity-score summary object}
\usage{
\method{print}{summary.causalmixgpd_ps_fit}(x, digits = 3, max_rows = 60, show_ess = FALSE, ...)
}
\arguments{
\item{x}{A \code{"summary.causalmixgpd_ps_fit"} object.}

\item{digits}{Number of digits to print in summary tables.}

\item{max_rows}{Maximum rows to print from the summary table.}

\item{show_ess}{Logical; if \code{TRUE}, include the \code{ess} column when present.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print a propensity-score summary object
}
