% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-kernel-registry.R
\name{register_kernel}
\alias{register_kernel}
\title{Register kernel}
\usage{
register_kernel(name, params, support, default_trans)
}
\arguments{
\item{name}{name.}

\item{params}{params.}

\item{support}{support.}

\item{default_trans}{default_trans.}
}
\value{
See details.
}
\description{
Register kernel.
}
\examples{
f <- getFromNamespace("register_kernel", "DPmixGPD")
f
\dontrun{
# Example usage (may require MCMC run and nimble compilation).
}

}
