% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-engine.R
\name{run_mcmc_nimble_gamma}
\alias{run_mcmc_nimble_gamma}
\title{Run mcmc nimble gamma}
\usage{
run_mcmc_nimble_gamma(spec, mcmc)
}
\arguments{
\item{spec}{spec.}

\item{mcmc}{mcmc.}
}
\value{
See details.
}
\description{
Run mcmc nimble gamma.
}
\examples{
f <- getFromNamespace("run_mcmc_nimble_gamma", "DPmixGPD")
f

}
\keyword{internal}
