% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated.R
\name{sim_survival_tail}
\alias{sim_survival_tail}
\title{Simulate survival tail data}
\usage{
sim_survival_tail(n = 250, seed = NULL)
}
\arguments{
\item{n}{Sample size.}

\item{seed}{Optional seed.}
}
\value{
A data.frame with time, status, and covariates.
}
\description{
Simulate survival tail data
}
