% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.causalmixgpd_causal_fit}
\alias{summary.causalmixgpd_causal_fit}
\title{Summarize a fitted causal model}
\usage{
\method{summary}{causalmixgpd_causal_fit}(object, ...)
}
\arguments{
\item{object}{A \code{"causalmixgpd_causal_fit"} object.}

\item{...}{Unused.}
}
\value{
The input object (invisibly).
}
\description{
\code{summary.causalmixgpd_causal_fit()} routes the user to the fitted PS
block, treated/control fits, and downstream prediction/effect methods.
}
\seealso{
\code{\link{print.causalmixgpd_causal_fit}},
\code{\link{predict.causalmixgpd_causal_fit}}, \code{\link{ate}},
\code{\link{qte}}, \code{\link{cate}}, \code{\link{cqte}}.
}
