% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-S3-Methods.R
\name{summary.causalmixgpd_causal_fit}
\alias{summary.causalmixgpd_causal_fit}
\title{Summarize a causal fit}
\usage{
\method{summary}{causalmixgpd_causal_fit}(object, ...)
}
\arguments{
\item{object}{A \code{"causalmixgpd_causal_fit"} object.}

\item{...}{Unused.}
}
\value{
The input object (invisibly).
}
\description{
Summarize a causal fit
}
