% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-model-methods.R
\name{summary.mixgpd_fit}
\alias{summary.mixgpd_fit}
\title{Summarize a MixGPD fit}
\usage{
\method{summary}{mixgpd_fit}(object, ...)
}
\arguments{
\item{object}{A fitted MixGPD object.}

\item{...}{Additional arguments (ignored).}
}
\value{
An object of class `summary.mixgpd_fit`.
}
\description{
Summarize a MixGPD fit
}
