% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-S3-Methods.R
\name{summary.mixgpd_fit}
\alias{summary.mixgpd_fit}
\title{Summarize a MixGPD fitted object}
\usage{
\method{summary}{mixgpd_fit}(object, pars = NULL, probs = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{object}{A fitted object of class \code{"mixgpd_fit"}.}

\item{pars}{Optional character vector of parameters to summarize. If NULL, summarize all (excluding v's).}

\item{probs}{Numeric vector of quantiles to report.}

\item{...}{Unused.}
}
\value{
An object of class \code{"mixgpd_summary"}.
}
\description{
Summarize a MixGPD fitted object
}
